api/project/requirements.txt
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
better-profanity==0.7.0
billiard==3.6.4.0
boto3==1.26.59
botocore==1.29.59
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jedi==0.19.1
jmespath==1.0.1
joblib==1.2.0
kombu==5.2.4
msgpack==1.0.4
numpy==1.24.2
packaging==23.2
parso==0.8.3
profanity-check==1.0.3
promise==2.3
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.22.0
psycopg2-binary==2.9.5
pudb==2023.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.16.1
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
rsa==4.9
Rx==1.6.3
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
service-identity==21.1.0
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
threadpoolctl==3.1.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
urwid==2.2.3
urwid-readline==0.13
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
